+++ src/org/simbrain/world/visionworld/pixelmatrix/EditableBufferedImagePixelMatrix.java
@@ -48,12 +48,12 @@ public final class EditableBufferedImagePixelMatrix
 
 
     /** {@inheritDoc} */
-    public double getHeight() {
+    public int getHeight() {
         return image.getHeight();
     }
 
     /** {@inheritDoc} */
-    public double getWidth() {
+    public int getWidth() {
         return image.getWidth();
     }
